Command Centers and fixes
Added: rudimentary command center
Fixed: events processed in reverse order
Fixed: packets had positions as integers, not floats

+++ src/io/github/alivety/conquerors/client/Model.java
@@ -60,7 +60,7 @@ public abstract class Model {
 		@Override
 		public Spatial build() {
 			final Node node = new Node(this.evt.spatialID);
-			for (final int[] shape : this.evt.form) {
+			for (final float[] shape : this.evt.form) {
 				/*
 				 * shape[0] = type (0=cube,1=sphere,2=cylindar)
 				 * shape[1-4] = RGBA color
