+++ src/io/github/alivety/conquerors/client/events/CreateModelEvent.java
@@ -6,10 +6,10 @@ import io.github.alivety.ppl.packet.Clientside;
 @Clientside
 public class CreateModelEvent extends Event {
 	public String spatialID;
-	public int[] position;
-	public int[][] form;
+	public float[] position;
+	public float[][] form;
 	
-	public CreateModelEvent(final String spatialID, final int[] position, final int[][] form) {
+	public CreateModelEvent(final String spatialID, final float[] position, final float[][] form) {
 		this.spatialID = spatialID;
 		this.position = position;
 		this.form = form;
